Dynamic background
Ugly and fake purple background is replaced by a more realistic
and dynamic star field background. Stars blink over the time
making them brighter or darker as time passes.

This still requires a lot of performance balancing, as I have
not tested yet how will this affect the performance of the game,
specially on mobile devices.